# add a local records queue in worker